polishing + update list endpoints, but has RS0017 errors
