+++ Consul/Interfaces/ITokenEndpoint.cs
@@ -43,6 +43,8 @@ namespace Consul
         Task<QueryResult<TokenEntry>> Read(string id, QueryOptions q, CancellationToken ct = default);
         Task<QueryResult<TokenEntry[]>> List(CancellationToken ct);
         Task<QueryResult<TokenEntry[]>> List();
+        Task<QueryResult<TokenEntry[]>> List(QueryOptions q, CancellationToken ct);
+        Task<QueryResult<TokenEntry[]>> List(QueryOptions q);
         Task<QueryResult<TokenEntry[]>> List(string policy, string role, string serviceName, string authMethod, CancellationToken ct);
         Task<QueryResult<TokenEntry[]>> List(string policy, string role, string serviceName, string authMethod);
         Task<QueryResult<TokenEntry[]>> List(string policy, string role, string serviceName, string authMethod, QueryOptions q, CancellationToken ct = default);
