+++ Consul/TemplatedPolicy.cs
@@ -1,0 +1,32 @@
+// -----------------------------------------------------------------------
+//  <copyright file="TemplatedPolicy.cs" company="PlayFab Inc">
+//    Copyright 2015 PlayFab Inc.
+//
+//    Licensed under the Apache License, Version 2.0 (the "License");
+//    you may not use this file except in compliance with the License.
+//    You may obtain a copy of the License at
+//
+//        http://www.apache.org/licenses/LICENSE-2.0
+//
+//    Unless required by applicable law or agreed to in writing, software
+//    distributed under the License is distributed on an "AS IS" BASIS,
+//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
+//  </copyright>
+// -----------------------------------------------------------------------
+
+using System.Collections.Generic;
+using Newtonsoft.Json;
+
+namespace Consul
+{
+    public class TemplatedPolicy
+    {
+        [JsonProperty("TemplateName")]
+        public string TemplateName { get; set; }
+
+        [JsonProperty("TemplateVariables")]
+        public Dictionary<string, object> TemplateVariables { get; set; }
+    }
+}
